requirements/test.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --resolver=backtracking test.in
#
astor==0.8.1
    # via flake8-simplify
attrs==23.1.0
    # via flake8-bugbear
bandit==1.7.5
    # via flake8-bandit
black==23.3.0
    # via -r test.in
click==8.1.3
    # via black
colorama==0.4.6
    # via
    #   bandit
    #   click
    #   pytest
coverage[toml]==7.2.7
    # via pytest-cov
flake8==6.0.0
    # via
    #   -r test.in
    #   flake8-bandit
    #   flake8-bugbear
    #   flake8-comprehensions
    #   flake8-docstrings
    #   flake8-html
    #   flake8-simplify
flake8-bandit==4.1.1
    # via -r test.in
flake8-bugbear==23.6.5
    # via -r test.in
flake8-comprehensions==3.13.0
    # via -r test.in
flake8-docstrings==1.7.0
    # via -r test.in
flake8-html==0.4.3
    # via -r test.in
flake8-plugin-utils==1.3.2
    # via flake8-pytest-style
flake8-pytest-style==1.7.2
    # via -r test.in
flake8-simplify==0.20.0
    # via -r test.in
gitdb==4.0.10
    # via gitpython
gitpython==3.1.31
    # via bandit
iniconfig==2.0.0
    # via pytest
invoke==2.1.3
    # via -r test.in
jinja2==3.1.2
    # via flake8-html
lxml==4.9.2
    # via -r test.in
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.3
    # via jinja2
mccabe==0.7.0
    # via flake8
mdurl==0.1.2
    # via markdown-it-py
mypy==1.4.1
    # via -r test.in
mypy-extensions==1.0.0
    # via
    #   black
    #   mypy
packaging==23.1
    # via
    #   black
    #   pytest
pathspec==0.11.1
    # via black
pbr==5.11.1
    # via stevedore
platformdirs==3.8.0
    # via black
pluggy==1.2.0
    # via pytest
py==1.11.0
    # via pytest-html
pycodestyle==2.10.0
    # via flake8
pydocstyle==6.3.0
    # via flake8-docstrings
pyflakes==3.0.1
    # via flake8
pygments==2.15.1
    # via
    #   flake8-html
    #   rich
pytest==7.4.0
    # via
    #   -r test.in
    #   pytest-cov
    #   pytest-html
    #   pytest-metadata
    #   pytest-mock
    #   pytest-randomly
    #   pytest-repeat
    #   pytest-timeout
pytest-cov==4.1.0
    # via -r test.in
pytest-html==3.2.0
    # via -r test.in
pytest-metadata==3.0.0
    # via
    #   -r test.in
    #   pytest-html
pytest-mock==3.11.1
    # via -r test.in
pytest-randomly==3.12.0
    # via -r test.in
pytest-repeat==0.9.1
    # via -r test.in
pytest-timeout==2.1.0
    # via -r test.in
pyyaml==6.0
    # via
    #   -r test.in
    #   bandit
rich==13.4.2
    # via bandit
smmap==5.0.0
    # via gitdb
snowballstemmer==2.2.0
    # via pydocstyle
stevedore==5.1.0
    # via bandit
types-beautifulsoup4==4.12.0.5
    # via -r test.in
types-click==7.1.8
    # via -r test.in
types-html5lib==1.1.11.14
    # via
    #   -r test.in
    #   types-beautifulsoup4
types-python-dateutil==2.8.19.13
    # via -r test.in
types-pyyaml==6.0.12.10
    # via -r test.in
types-requests==2.31.0.1
    # via -r test.in
types-urllib3==1.26.25.13
    # via
    #   -r test.in
    #   types-requests
typing-extensions==4.6.3
    # via mypy
